IDEADEV-22547 Class Name Code Completion: on completing one sybol is removed

+++ codeInsight/impl/com/intellij/codeInsight/completion/DefaultInsertHandler.java
@@ -694,7 +694,9 @@ public class DefaultInsertHandler implements InsertHandler,Cloneable {
       }
     }
 
-    document.deleteString(toDelete.getStartOffset(), toDelete.getEndOffset());
+    if (toDelete.isValid()) {
+      document.deleteString(toDelete.getStartOffset(), toDelete.getEndOffset());
+    }
 
     return newStartOffset;
   }
